refactor: Rename UseLogin to UseLogin2 for case-change in Git
